Adding administration controll endpoints
UserController:
-indexAdmins
-deleteAdmin

VCardController:
-indexVCards
-updateVCard
-deleteVCard

"Administrators will also be able to manage other administrators – view the list of administrators
and add or remove administrator accounts – and manage the vCards, where they can view, filter,
and search the list of vCards. For each vCard they can access the phone number; name, e-mail and
photo of the owner; balance and debit limit (max_debit) of the vCard and whether the vCard is
blocked. Also, the administrator can block or unblock any vCard, change the debit limit
(max_debit) of any vCard and delete vCards that have a balance of zero – if the vCard has no
associated transactions, it is deleted from the database, if the vCard has associated transactions,
then the vCard and all associated transactions are "soft deleted"